Put functionality related to command-line args in own cpp file
